refactor(123): extract OrganizationDto + BrandingDto to Services/Shared/Organization/
T020-T022 — first DTO extraction applying the pattern documented in
contracts/shared-dto-extraction-pattern.md.

OrganizationDto is consumed by PublishParticipantDialog (user-facing)
as a [Parameter] type as well as by 8 admin pages. BrandingDto is
extracted as a pair with it (OrganizationDto.Branding references
BrandingDto — they're a natural unit).

Audit discovery: UserDto turned out to be admin-only after Layer-1 grep
(only consumed by admin user-management pages). Plan had assumed it was
shared; correction recorded — UserDto stays co-located with the admin
interface and will move to Services/Admin/ when the interface file
itself moves in Phase 3.

IOrganizationReadService creation deferred: PublishParticipantDialog
receives OrganizationDto via [Parameter], not via service injection.
No current consumer would inject IOrganizationReadService, so creating
it would be YAGNI. If a future user-facing page needs the read
operation as an injected call, add it then.

Build: dotnet build src/Apps/Sorcha.UI/Sorcha.UI.sln passes, zero errors.
